backend/requirements.txt
# Backend Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
reportlab==4.0.7
Pillow==10.1.0
matplotlib==3.8.2
